+++ changelog.md
@@ -33,7 +33,7 @@ This changelog is based on [Keep A Changelog](https://keepachangelog.com/en/1.1.
 
 ## Added
 
-* Added `smart-contracts` command to the node with first command `init-governance`.
+* Added `smart-contracts` command to the node with sub-commands `init-governance` and `get-scripts`.
 
 # v1.3.0